requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncpg]
pydantic
python-jose[cryptography]
python-dotenv
openai
psycopg2-binary
